Move invocation to a trait
